Bien documentado

+++ Calculadora/biblioteca.c
@@ -1,7 +1,13 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include "biblioteca.h"
-
+/** \brief Suma dos numeros enteros
+ *
+ * \param int primer operador de la suma
+ * \param int segundo operador de la suma
+ * \return int resultado de la suma de dos enteros
+ *
+ */
 int sumar(int numeroUno, int numeroDos)
 {
     int suma;
@@ -11,6 +17,14 @@ int sumar(int numeroUno, int numeroDos)
     return suma;
 }
 
+
+/** \brief Resta dos numeros enteros
+ *
+ * \param int primero operador de la resta
+ * \param int segundo operador de la resta
+ * \return int resultado de la resta entre dos numeros enteros
+ *
+ */
 int restar (int numeroUno, int numeroDos)
 {
     int resta;
@@ -20,6 +34,14 @@ int restar (int numeroUno, int numeroDos)
     return resta;
 }
 
+
+/** \brief Divide dos numeros enteros
+ *
+ * \param int Dividendo
+ * \param int Divisor
+ * \return float cociente de la division entre dos enteros
+ *
+ */
 float dividir(int numeroUno, int numeroDos)
 {
     float division;
@@ -29,6 +51,14 @@ float dividir(int numeroUno, int numeroDos)
     return division;
 }
 
+
+/** \brief Multiplica dos enteros
+ *
+ * \param int primer operador de la multiplicacion
+ * \param int segundo operador de la multiplicacion
+ * \return resultado de la multiplicacion entre dos enteros
+ *
+ */
 int multiplicar (int numeroUno, int numeroDos)
 {
     int multiplicacion;
@@ -38,7 +68,15 @@ int multiplicar (int numeroUno, int numeroDos)
     return multiplicacion;
 }
 
-int factorialA (int numeroUno)
+
+
+/** \brief muestra el factorial de un natural
+ *
+ * \param int Numero Natural del cual se quiere sacar el factorial
+ * \return int factorial del numero entero ingresado
+ *
+ */
+long int factorialA (int numeroUno)
 {
     int factorial=1;
     int i;
@@ -50,7 +88,16 @@ int factorialA (int numeroUno)
       return factorial;
 }
 
-int factorialB (int numeroDos)
+
+
+
+/** \brief muestra el factorial de un natural
+ *
+ * \param int Numero Natural del cual se quiere sacar el factorial
+ * \return int factorial del numero entero ingresado
+ *
+ */
+long int factorialB (int numeroDos)
 {
     int factorial=1;
     int i;
@@ -61,6 +108,15 @@ int factorialB (int numeroDos)
 
       return factorial;
 }
+
+
+
+/** \brief Pide un numero entero
+ *
+ * \param char [] Texto que le pida al usuario que ingrese un numero
+ * \return int numero entero que ingresa el usuario
+ *
+ */
 
 int pedirEntero (char texto[])
 {
