+++ Calculadora/biblioteca.h
@@ -1,73 +1,27 @@
 
-/** \brief Suma dos numeros enteros
- *
- * \param int primer operador de la suma
- * \param int segundo operador de la suma
- * \return int resultado de la suma de dos enteros
- *
- */
+
 int sumar (int , int);
 
 
-/** \brief Resta dos numeros enteros
- *
- * \param int primero operador de la resta
- * \param int segundo operador de la resta
- * \return int resultado de la resta entre dos numeros enteros
- *
- */
+
 int restar (int,int);
 
 
-
-
-/** \brief Divide dos numeros enteros
- *
- * \param int Dividendo
- * \param int Divisor
- * \return float cociente de la division entre dos enteros
- *
- */
 float dividir (int,int);
 
 
-
-/** \brief Multiplica dos enteros
- *
- * \param int primer operador de la multiplicacion
- * \param int segundo operador de la multiplicacion
- * \return resultado de la multiplicacion entre dos enteros
- *
- */
 int multiplicar (int, int);
 
 
 
-/** \brief muestra el factorial de un natural
- *
- * \param int Numero Natural del cual se quiere sacar el factorial
- * \return int factorial del numero entero ingresado
- *
- */
-int factorialA (int);
+
+long int factorialA (int);
 
 
 
-/** \brief muestra el factorial de un natural
- *
- * \param int Numero Natural del cual se quiere sacar el factorial
- * \return int factorial del numero entero ingresado
- *
- */
-int factorialB (int);
+long int factorialB (int);
 
 
 
-/** \brief Pide un numero entero
- *
- * \param char [] Texto que le pida al usuario que ingrese un numero
- * \return int numero entero que ingresa el usuario
- *
- */
 
 int pedirEntero (char[]);
